+++ src/test/java/com/redhat/demos/quarkusretailstore/inventory/InventoryResourceTest.java
@@ -44,8 +44,8 @@ public class InventoryResourceTest {
 
         with().body(JsonUtil.toJson(new InventoryJson(
                         new ProductMaster(skuId, "A new product description"),
-                        BigDecimal.valueOf(19.99),
-                        BigDecimal.valueOf(24.99),
+                        19.99,
+                        24.99,
                         1,
                         9,
                         0,
@@ -67,8 +67,8 @@ public class InventoryResourceTest {
 
         System.out.println(JsonUtil.toJson(new InventoryJson(
                 new ProductMaster(skuId, "A product description"),
-                BigDecimal.valueOf(19.99),
-                BigDecimal.valueOf(24.99),
+                19.99,
+                24.99,
                 1,
                 9,
                 0,
@@ -80,8 +80,8 @@ public class InventoryResourceTest {
 
         with().body(JsonUtil.toJson(new InventoryJson(
                         new ProductMaster(skuId, "A product description"),
-                        BigDecimal.valueOf(19.99),
-                        BigDecimal.valueOf(24.99),
+                        19.99,
+                      24.99,
                         1,
                         9,
                         0,
